AI chat App/requirements.txt
streamlit
pdfplumber
langchain
openai
huggingface_hub
sentence-transformers
faiss-cpu
fastapi
uvicorn